refactor: move updates checks to its own module
